Added a file to /docs
This is primarily to test if the /docs directory can be used in GitHub Pages rather than the root directory in the master branch.  It appears greyed out in the web UI so I'm trying this to see if Github will allow this option once something is present in /docs.